+++ src/base/event/dispatch.h
@@ -40,8 +40,8 @@ struct mm_event_dispatch_attr
 	uint32_t dispatch_queue_size;
 	uint32_t listener_queue_size;
 	/* Spinning parameters. */
-	uint16_t lock_spin_limit;
-	uint16_t poll_spin_limit;
+	uint32_t lock_spin_limit;
+	uint32_t poll_spin_limit;
 
 	/* Individual listener parameters. */
 	struct mm_event_dispatch_listener_attr *listeners_attr;
@@ -55,8 +55,8 @@ struct mm_event_dispatch
 	mm_thread_t nlisteners;
 
 	/* Spinning parameters. */
-	uint16_t lock_spin_limit;
-	uint16_t poll_spin_limit;
+	uint32_t lock_spin_limit;
+	uint32_t poll_spin_limit;
 
 	/* Asynchronous post queue. */
 	struct mm_ring_mpmc *async_queue;
@@ -72,7 +72,7 @@ struct mm_event_dispatch
 	mm_regular_lock_t poller_lock CACHE_ALIGN;
 
 	/* Counter for poller thread busy waiting. */
-	uint16_t poll_spin_count;
+	uint32_t poll_spin_count;
 
 #if ENABLE_EVENT_SINK_LOCK
 	/* A coarse-grained lock that protects event sinks from
@@ -101,9 +101,9 @@ void NONNULL(1)
 mm_event_dispatch_attr_setlistenerqueuesize(struct mm_event_dispatch_attr *attr, uint32_t size);
 
 void NONNULL(1)
-mm_event_dispatch_attr_setlockspinlimit(struct mm_event_dispatch_attr *attr, uint16_t value);
+mm_event_dispatch_attr_setlockspinlimit(struct mm_event_dispatch_attr *attr, uint32_t value);
 void NONNULL(1)
-mm_event_dispatch_attr_setpollspinlimit(struct mm_event_dispatch_attr *attr, uint16_t value);
+mm_event_dispatch_attr_setpollspinlimit(struct mm_event_dispatch_attr *attr, uint32_t value);
 
 void NONNULL(1, 3)
 mm_event_dispatch_attr_setlistenerstrand(struct mm_event_dispatch_attr *attr, mm_thread_t n, struct mm_strand *strand);
